backend/requirements.txt
# Python Backend for SAGA Store - LIMS Integration
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.4
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
python-multipart==0.0.6

# For LIMS integration (add your company's LIMS client library here)
# Example: lims-client==1.0.0
